+++ app/views/provider/home.php
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Customer</title>
+    <title>Proveedores</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.7.2/css/all.min.css"
@@ -39,8 +39,10 @@
                         <th scope="row"><?= $provider->provider_id ?></th>
                         <td><?= $provider->name ?></td>
                         <td>
-                            <i class="fa-solid fa-user-pen"></i>
-                            <a href="<?= base_url() ?>provider/deleteProvider/<?= $provider->provider_id ?>">
+                            <a href="<?= base_url() ?>provider/editProvider/<?= $provider->provider_id ?>">
+                                <i class="fa-solid fa-user-pen"></i>
+                            </a>
+                            <a href="<?= base_url() ?>provider/deleteProvider/<?= $provider->provider_id ?>" onclick="return confirm('¿Está seguro de que desea eliminar este proveedor?')">
                                 <i class="fa-solid fa-trash"></i>
                             </a>
                             <a href="<?= base_url() ?>provider/show/<?= $provider->provider_id ?>">
